+++ Common.Services/MassConverter.cs
@@ -4,15 +4,15 @@ namespace Common.Services
 {
     public class MassConverter : IMassConverter
     {
-        public Dictionary<MassUnit, double> Conversions = new Dictionary<MassUnit, double>()
+        public Dictionary<MassUnit, decimal> Conversions = new Dictionary<MassUnit, decimal>()
         {
             { MassUnit.kg, 1 },
             { MassUnit.gr, 1000 }
         };
 
-        public double Convert(double fromValue, MassUnit fromUnit, MassUnit toUnit)
+        public decimal Convert(decimal fromValue, MassUnit fromUnit, MassUnit toUnit)
         {
-            double workingValue;
+            decimal workingValue;
 
             if (fromUnit == MassUnit.kg)
                 workingValue = fromValue;
